requirements.txt
asgiref==3.8.1
Django==5.0.3
sqlparse==0.4.4
firebase-admin==6.5.0
bcrypt==4.1.2
pandas==2.2.2
python-dotenv==1.0.1
elasticsearch==8.13.0
elasticsearch-dsl==8.13.0